requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
coverage==6.1.2
idna==3.3
Pillow==8.4.0
pony==0.7.14
psycopg2==2.9.2
requests==2.26.0
state==0.1.1.dev0
urllib3==1.26.7
vk-api==11.9.6
